chore: refactor pipeline forwarders
