+++ rust/numaflow-core/src/pipeline/forwarder/source_forwarder.rs
@@ -1,9 +1,39 @@
-use tokio_util::sync::CancellationToken;
-
-use crate::error;
+use crate::config::components::metrics::MetricsConfig;
+use crate::config::components::source::{GeneratorConfig, SourceConfig, SourceType};
+use crate::config::is_mono_vertex;
+use crate::config::pipeline::isb::BufferFullStrategy::RetryUntilSuccess;
+use crate::config::pipeline::isb::{BufferWriterConfig, Stream};
+use crate::config::pipeline::{
+    PipelineConfig, SourceVtxConfig, ToVertexConfig, VertexConfig, VertexType, isb,
+};
 use crate::error::Error;
-use crate::pipeline::isb::jetstream::writer::JetstreamWriter;
+use crate::metrics::{
+    ComponentHealthChecks, LagReader, MetricsState, PendingReaderTasks, PipelineComponents,
+    WatermarkFetcherState,
+};
+use crate::pipeline::PipelineContext;
+use crate::pipeline::forwarder::start_forwarder;
+use crate::pipeline::isb::jetstream::writer::{ISBWriterComponents, JetstreamWriter};
+use crate::shared::create_components;
+use crate::shared::metrics::start_metrics_server;
 use crate::source::Source;
+use crate::tracker::TrackerHandle;
+use crate::transformer::Transformer;
+use crate::typ::{
+    NumaflowTypeConfig, WithInMemoryRateLimiter, WithRedisRateLimiter, WithoutRateLimiter,
+    build_in_memory_rate_limiter_config, build_redis_rate_limiter_config,
+    should_use_redis_rate_limiter,
+};
+use crate::watermark::WatermarkHandle;
+use crate::watermark::source::SourceWatermarkHandle;
+use crate::{error, shared};
+use async_nats::jetstream;
+use async_nats::jetstream::{Context, consumer, stream};
+use serving::callback::CallbackHandler;
+use std::time::Duration;
+use tokio_stream::StreamExt;
+use tokio_util::sync::CancellationToken;
+use tracing::info;
 
 /// Source forwarder is the orchestrator which starts streaming source, a transformer, and an isb writer
 /// and manages the lifecycle of these components.
@@ -47,8 +77,152 @@ impl<C: crate::typ::NumaflowTypeConfig> SourceForwarder<C> {
     }
 }
 
+pub(crate) async fn start_source_forwarder(
+    cln_token: CancellationToken,
+    js_context: Context,
+    config: PipelineConfig,
+    source_config: SourceVtxConfig,
+    source_watermark_handle: Option<SourceWatermarkHandle>,
+) -> error::Result<()> {
+    let serving_callback_handler = if let Some(cb_cfg) = &config.callback_config {
+        Some(
+            CallbackHandler::new(
+                config.vertex_name.to_string(),
+                js_context.clone(),
+                cb_cfg.callback_store,
+                cb_cfg.callback_concurrency,
+            )
+            .await,
+        )
+    } else {
+        None
+    };
+
+    let tracker_handle = TrackerHandle::new(serving_callback_handler);
+
+    let context = PipelineContext {
+        cln_token: cln_token.clone(),
+        js_context: &js_context,
+        config: &config,
+        tracker_handle: tracker_handle.clone(),
+    };
+
+    let writer_components = ISBWriterComponents::new(
+        source_watermark_handle.clone().map(WatermarkHandle::Source),
+        &context,
+    );
+
+    let buffer_writer = JetstreamWriter::new(writer_components);
+    let transformer = create_components::create_transformer(
+        config.batch_size,
+        config.graceful_shutdown_time,
+        source_config.transformer_config.clone(),
+        tracker_handle.clone(),
+        cln_token.clone(),
+    )
+    .await?;
+
+    // Apply rate limiting dispatch pattern similar to other forwarders
+    if let Some(rate_limit_config) = &config.rate_limit {
+        if should_use_redis_rate_limiter(rate_limit_config) {
+            let redis_config =
+                build_redis_rate_limiter_config(rate_limit_config, cln_token.clone()).await?;
+
+            run_source_forwarder::<WithRedisRateLimiter>(
+                &context,
+                &source_config,
+                transformer,
+                source_watermark_handle,
+                buffer_writer,
+                Some(redis_config.throttling_config),
+            )
+            .await?
+        } else {
+            let in_mem_config =
+                build_in_memory_rate_limiter_config(rate_limit_config, cln_token.clone()).await?;
+
+            run_source_forwarder::<WithInMemoryRateLimiter>(
+                &context,
+                &source_config,
+                transformer,
+                source_watermark_handle,
+                buffer_writer,
+                Some(in_mem_config.throttling_config),
+            )
+            .await?
+        }
+    } else {
+        run_source_forwarder::<WithoutRateLimiter>(
+            &context,
+            &source_config,
+            transformer,
+            source_watermark_handle,
+            buffer_writer,
+            None,
+        )
+        .await?
+    };
+
+    Ok(())
+}
+
+/// Starts source forwarder.
+async fn run_source_forwarder<C: NumaflowTypeConfig>(
+    context: &PipelineContext<'_>,
+    source_config: &SourceVtxConfig,
+    transformer: Option<Transformer>,
+    source_watermark_handle: Option<SourceWatermarkHandle>,
+    buffer_writer: JetstreamWriter,
+    rate_limiter: Option<C::RateLimiter>,
+) -> error::Result<()> {
+    let source = create_components::create_source::<C>(
+        context.config.batch_size,
+        context.config.read_timeout,
+        &source_config.source_config,
+        context.tracker_handle.clone(),
+        transformer,
+        source_watermark_handle.clone(),
+        context.cln_token.clone(),
+        rate_limiter,
+    )
+    .await?;
+
+    // only check the pending and lag for source for pod_id = 0
+    let _pending_reader_handle: Option<PendingReaderTasks> = if context.config.replica == 0 {
+        let pending_reader = shared::metrics::create_pending_reader::<C>(
+            &context.config.metrics_config,
+            LagReader::Source(Box::new(source.clone())),
+        )
+        .await;
+        info!("Started pending reader");
+        Some(pending_reader.start(is_mono_vertex()).await)
+    } else {
+        None
+    };
+
+    start_metrics_server::<C>(
+        context.config.metrics_config.clone(),
+        MetricsState {
+            health_checks: ComponentHealthChecks::Pipeline(Box::new(PipelineComponents::Source(
+                Box::new(source.clone()),
+            ))),
+            watermark_fetcher_state: source_watermark_handle.map(|handle| WatermarkFetcherState {
+                watermark_handle: WatermarkHandle::Source(handle),
+                partition_count: 1, // Source vertices always have partition count = 1
+            }),
+        },
+    )
+    .await;
+
+    let forwarder = SourceForwarder::<C>::new(source, buffer_writer);
+
+    forwarder.start(context.cln_token.clone()).await
+}
+
 #[cfg(test)]
 mod tests {
+    use super::*;
+
     use std::collections::HashSet;
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::time::Duration;
@@ -319,5 +493,146 @@ mod tests {
         src_shutdown_tx.send(()).unwrap();
         source_handle.await.unwrap();
         transformer_handle.await.unwrap();
+    }
+
+    // e2e test for source forwarder, reads from generator and writes to
+    // multi-partitioned buffer.
+    #[cfg(feature = "nats-tests")]
+    #[tokio::test]
+    async fn test_forwarder_for_source_vertex() {
+        // Unique names for the streams we use in this test
+        let streams = vec![
+            Stream::new("default-test-forwarder-for-source-vertex-out-0", "test", 0),
+            Stream::new("default-test-forwarder-for-source-vertex-out-1", "test", 1),
+            Stream::new("default-test-forwarder-for-source-vertex-out-2", "test", 2),
+            Stream::new("default-test-forwarder-for-source-vertex-out-3", "test", 3),
+            Stream::new("default-test-forwarder-for-source-vertex-out-4", "test", 4),
+        ];
+
+        let js_url = "localhost:4222";
+        let client = async_nats::connect(js_url).await.unwrap();
+        let context = jetstream::new(client);
+
+        let mut consumers = vec![];
+        // Create streams to which the generator source vertex we create later will forward
+        // messages to. The consumers created for the corresponding streams will be used to ensure
+        // that messages were actually written to the streams.
+        for stream in &streams {
+            // Delete stream if it exists
+            let _ = context.delete_stream(stream.name).await;
+            let _stream = context
+                .get_or_create_stream(stream::Config {
+                    name: stream.name.to_string(),
+                    subjects: vec![stream.name.into()],
+                    max_message_size: 64 * 1024,
+                    max_messages: 10000,
+                    ..Default::default()
+                })
+                .await
+                .unwrap();
+
+            let c: consumer::PullConsumer = context
+                .create_consumer_on_stream(
+                    consumer::pull::Config {
+                        name: Some(stream.to_string()),
+                        ack_policy: consumer::AckPolicy::Explicit,
+                        ..Default::default()
+                    },
+                    stream.name,
+                )
+                .await
+                .unwrap();
+            consumers.push((stream.to_string(), c));
+        }
+
+        let pipeline_config = PipelineConfig {
+            pipeline_name: "simple-pipeline",
+            vertex_name: "in",
+            replica: 0,
+            batch_size: 1000,
+            writer_concurrency: 30000,
+            read_timeout: Duration::from_secs(1),
+            js_client_config: isb::jetstream::ClientConfig {
+                url: "localhost:4222".to_string(),
+                user: None,
+                password: None,
+            },
+            from_vertex_config: vec![],
+            to_vertex_config: vec![ToVertexConfig {
+                name: "out",
+                partitions: 5,
+                writer_config: BufferWriterConfig {
+                    streams: streams.clone(),
+                    max_length: 30000,
+                    usage_limit: 0.8,
+                    buffer_full_strategy: RetryUntilSuccess,
+                },
+                conditions: None,
+                to_vertex_type: VertexType::Sink,
+            }],
+            vertex_type: VertexType::Source,
+            vertex_config: VertexConfig::Source(SourceVtxConfig {
+                source_config: SourceConfig {
+                    read_ahead: false,
+                    source_type: crate::config::components::source::SourceType::Generator(
+                        GeneratorConfig {
+                            rpu: 10,
+                            content: bytes::Bytes::new(),
+                            duration: Duration::from_secs(1),
+                            value: None,
+                            key_count: 0,
+                            msg_size_bytes: 300,
+                            jitter: Duration::from_millis(0),
+                        },
+                    ),
+                },
+                transformer_config: None,
+            }),
+            metrics_config: MetricsConfig {
+                metrics_server_listen_port: 2469,
+                lag_check_interval_in_secs: 5,
+                lag_refresh_interval_in_secs: 3,
+                lookback_window_in_secs: 120,
+            },
+            ..Default::default()
+        };
+
+        let cancellation_token = CancellationToken::new();
+        let forwarder_task = tokio::spawn({
+            let cancellation_token = cancellation_token.clone();
+            async move {
+                start_forwarder(cancellation_token, pipeline_config)
+                    .await
+                    .unwrap();
+            }
+        });
+
+        // Wait for a few messages to be forwarded
+        tokio::time::sleep(Duration::from_secs(2)).await;
+        cancellation_token.cancel();
+        forwarder_task.await.unwrap();
+
+        for (stream_name, stream_consumer) in consumers {
+            let messages: Vec<jetstream::Message> = stream_consumer
+                .batch()
+                .max_messages(10)
+                .expires(Duration::from_millis(50))
+                .messages()
+                .await
+                .unwrap()
+                .map(|msg| msg.unwrap())
+                .collect()
+                .await;
+            assert!(
+                !messages.is_empty(),
+                "Stream {} is expected to have messages",
+                stream_name
+            );
+        }
+
+        // Delete all streams created in this test
+        for stream in streams {
+            context.delete_stream(stream.name).await.unwrap();
+        }
     }
 }
